Move pagination config to a separate module
